Continue running when EnumerateDeviceExtensionProperties call result is VK_INCOMPLETE
When getting device extension properities from the driver, the function
EnumerateDeviceExtensionProperties() will be called 2 times. In the first call,
the function will return the number of extensions available, N, whereas in the
second, it will get the extension names.
Due to the fact that the last extension name given from the driver is a NULL
string, the second call will return N - 1 extensions, resulting in
VK_INCOMPLETE.
This commit fixes the capture support on Vulkan for iMX8 targets.