Pass memory checking when replaying captures
Due to the fact that iMX8 target fails buffer size checking when loading the
buffers from the capture file, the replay feauture is not working.
This commit adds a round up for the allocation size to allow creation of
the buffers, by creating a dummy buffer which has the initial allocation size.
If this buffer pass the checkings, then the allocation size is correctly set,
else it will become equal to the demanded size. At the end, the dummy buffer
is freed.